refactorred functions
