+++ packages/builder/src/compile.ts
@@ -4,7 +4,13 @@ import { Compiler } from '@tsx-air/types';
 // @ts-ignore
 import { createReporter } from './reporter';
 import isString from 'lodash/isString';
+import * as runtime from '@tsx-air/runtime';
+import { Script } from 'vm';
+import { join } from 'path';
 
+/**
+ * Compiles a list of files and their imports (excluding node-modules)
+ */
 export function compile(fileNames: string[], compiler: Compiler, outDir: string, report?: string | boolean) {
     const program = ts.createProgram(fileNames,
         { ...compilerOptions, outDir, noEmit: !outDir, sourceMap: true });
@@ -35,4 +41,30 @@ export function compile(fileNames: string[], compiler: Compiler, outDir: string,
         );
     }
     return allDiagnostics;
+}
+
+
+export function compileAndEval(content: string, compiler: Compiler, importPath: string = '', executionContext = {}) {
+    const out = ts.transpileModule(content, {
+        compilerOptions: {
+            ...compilerOptions,
+            module: ts.ModuleKind.CommonJS
+        },
+        transformers: compiler.transformers,
+        fileName: 'compiled.jsx'
+    }).outputText;
+
+    const exports = {};
+    const _require = (path: string) => {
+        switch (path) {
+            case '@tsx-air/runtime':
+                return runtime;
+            default:
+                return require(join(importPath, path));
+        }
+    };
+    const script = new Script(out);
+    const context = { ...runtime, require: _require, exports, ...executionContext };
+    script.runInNewContext(context);
+    return exports as Record<string, any>;
 }
