+++ packages/compilers/src/ast-based-compiler/component/fragment/to.string.test.ts
@@ -1,41 +1,43 @@
-import { basicPatterns } from '../../../test.helpers';
-import { CompDefinition, evalAst } from '@tsx-air/compiler-utils';
-import { generateToString } from './to.string';
+import { readFixture } from '../../../test.helpers';
 import { chaiPlugin } from '@tsx-air/testing';
 import { expect, use } from 'chai';
-import { asFunction } from '../function';
-import { parseFragments, FragmentData } from './jsx.fragment';
-import { identity } from 'lodash';
-import * as runtime from '@tsx-air/runtime';
+import { compileAndEval } from '@tsx-air/builder';
+import { astBasedCompiler } from '@tsx-air/compilers';
+import { Component, Runtime } from '@tsx-air/runtime';
+import { JSDOM } from 'jsdom';
 use(chaiPlugin);
 
 describe('generateToString', () => {
-    afterEach(runtime.reset);
-    const toStringOf = (comp: CompDefinition, props: any, scope = {}) => {
-        const frag = parseFragments(comp).next().value as FragmentData;
-        const asFunc = evalAst(asFunction(generateToString(frag))) as () => any;
-        return () => asFunc.apply({ $rt: runtime.getInstance(), stores: { $props: props }, ...scope, unique: identity });
+    let comps: {
+        Static: typeof Component;
+        PropsOnly: typeof Component;
+        StateOnly: typeof Component;
+        ProsAndState: typeof Component;
+        NestedStateless: typeof Component;
+        EventListener: typeof Component;
+        DynamicAttributes: typeof Component;
+        DynamicAttributesSelfClosing: typeof Component;
+        WithVolatile: typeof Component;
+    };
+    before(() => {
+        comps = compileAndEval(readFixture(`basic.patterns.tsx`), astBasedCompiler) as any;
+    });
+    const toStringOf = (comp: typeof Component, props: any) => {
+        const { window } = new JSDOM(`<!DOCTYPE html><html><body></body></html>`);
+        const instance = new comp('', undefined, props, new Runtime(window, () => void (0)));
+        return instance.toString();
     };
     it('generates a toString method based on the used props and state', () => {
-        const comps = basicPatterns();
-
-        expect(toStringOf(comps.Static, {})(), 'Static').to.equal('<div></div>');
-        expect(toStringOf(comps.PropsOnly, { 'props.a': 'a', 'props.b': 'b' })(), 'PropsOnly')
-            .to.equal(`<div><!--X-->a<!--X--><!--X-->b<!--X--></div>`);
-        expect(toStringOf(comps.StateOnly, {'store1.a':1, 'store1.b':2})(), 'StateOnly')
-            .to.equal(`<div><!--X-->1<!--X--><!--X-->2<!--X--></div>`);
-        expect(toStringOf(comps.DynamicAttributes, { 'props.a': 1 })(), 'DynamicAttributes')
-            .to.equal(`<div dir="ltr" lang="1" x-da="!"><span></span></div>`);
-        expect(toStringOf(comps.DynamicAttributesSelfClosing, { 'props.a': 2 })(), 'DynamicAttributesSelfClosing')
-            .to.equal(`<div dir="ltr" lang="2" x-da="!"></div>`);
-        expect(toStringOf(comps.WithVolatile, {d:'volatile'})(), 'WithVolatile')
-            .to.equal(`<div><!--X-->volatile<!--X--></div>`);
+        expect(toStringOf(comps.Static, {}), 'Static').to.equal('<div></div>');
+        expect(toStringOf(comps.PropsOnly, { 'a': 'a', 'b': 'b' }), 'PropsOnly')
+            .to.equal(`<div><!--0X0-->a<!--0X0--><!--0X1-->b<!--0X1--></div>`);
+        expect(toStringOf(comps.DynamicAttributes, { 'a': 1 }), 'DynamicAttributes')
+            .to.equal(`<div dir="ltr" lang="1" x-da="0"><span></span></div>`);
+        expect(toStringOf(comps.DynamicAttributesSelfClosing, { 'a': 2 }), 'DynamicAttributesSelfClosing')
+            .to.equal(`<div dir="ltr" lang="2" x-da="0"></div>`);
     });
 
     it(`removes event listeners`, () => {
-        const { EventListener } = basicPatterns();
-        const withEvent = toStringOf(EventListener, {});
-
-        expect(withEvent()).to.eql(`<div x-da="!"></div>`);
+        expect(toStringOf(comps.EventListener, {})).to.eql(`<div x-da="0"></div>`);
     });
 });
