+++ packages/compilers/src/ast-based-compiler/component/functions/index.ts
@@ -1,0 +1,2 @@
+export * from './function';
+export { swapVirtualElements } from './swappers';
