+++ packages/compilers/src/index.test.ts
@@ -2,12 +2,8 @@ import { expect } from 'chai';
 import { transformerCompilers } from '.';
 import { packagePath } from '@tsx-air/utils/packages';
 import { readFileSync } from 'fs';
-import ts from 'typescript';
-import { compilerOptions } from '@tsx-air/compiler-utils';
-import { Script } from 'vm';
-import * as runtime from '@tsx-air/runtime';
-import { browserifyFiles } from '@tsx-air/testing';
 import { testRuntimeApi } from '@tsx-air/runtime/src/runtime/runtime.test.suite';
+import { compileAndEval } from '@tsx-air/builder/src';
 
 describe('compilers', () => {
     it('each compiler should have a unique name', () => {
@@ -24,28 +20,18 @@ describe('compilers', () => {
             let Parent: any;
             let Child: any;
             before(async () => {
-                await browserifyFiles(
-                    packagePath('@tsx-air/runtime', 'fixtures'),
-                    packagePath('@tsx-air/compilers', 'tmp'),
-                    compiler,
-                    'runtime.fixture.tsx',
-                    'out.js',
+                // await buildTestFiles(
+                //     packagePath('@tsx-air/runtime', 'fixtures'),
+                //     packagePath('@tsx-air/compilers', 'tmp'),
+                //     compiler,
+                //     'runtime.fixture.tsx',
+                //     'out.js',
+                // );
+
+                const exports = compileAndEval(
+                    readFileSync(packagePath('@tsx-air/runtime', 'fixtures', 'runtime.fixture.tsx'), { encoding: 'utf8' }),
+                    compiler
                 );
-                const src = readFileSync(packagePath('@tsx-air/runtime', 'fixtures', 'runtime.fixture.tsx'), { encoding: 'utf8' });
-                const out = ts.transpileModule(src, {
-                    compilerOptions: {
-                        ...compilerOptions,
-                        module: ts.ModuleKind.CommonJS
-                    },
-                    transformers: compiler.transformers,
-                    fileName: 'compiled.jsx'
-                }).outputText;
-               
-                const exports = { Parent: null, Child: null };
-                const require = () => runtime;
-                const script = new Script(out);
-                const context = {...runtime, require, exports, $rt:runtime.getInstance};
-                script.runInNewContext(context);
                 Parent = exports.Parent;
                 Child = exports.Child;
             });
