+++ packages/examples/src/examples/index.ts
@@ -1,8 +1,8 @@
 export const shouldBeCompiled = [
-    // '01.stateless-parent-child',
-    // '02.stateful',
-    // '03.thumb',
-    // '04.zoom',
+    '01.stateless-parent-child',
+    '02.stateful',
+    '03.thumb',
+    '04.zoom',
     '05.dynamic.gallery',
     // '06.factory'
 ];
