+++ packages/framework/src/api/lifecycle.ts
@@ -1,6 +1,13 @@
-export function when(_predicate: any | any[], _action: () => void) {/* */}
-export function memo(_predicate: any | any[], _action: () => any) {/* */}
-export async function invalidate(): Promise<void> {/* */}
-export function afterMount(_cb: (rootRef: HTMLElement) => void) {/* */}
-export function beforeUnmount(_cb: (rootRef: HTMLElement) => void) {/* */}
-export function afterDomUpdate(_predicate: any | any[], _cb: () => void) {/* */}
+export function when(_action: () => void): void;
+export function when(_predicate: any | any[], _action: () => void): void;
+export function when(_a1: any | any[], _a2?: () => void) {/* */ }
+
+export function memo(_action: () => any): any;
+export function memo(_predicate: any | any[], _action: () => any): any;
+export function memo(_a1: any | any[], _a2?: () => void): any {/* */ }
+
+export function invalidate():void {/* */ }
+
+export function afterMount(_cb: (rootRef: HTMLElement) => void|(()=>void)) {/* */ }
+
+export function afterDomUpdate(_predicate: any | any[], _cb: () => void) {/* */ }
