+++ packages/testing/src/test.using.examples.ts
@@ -1,12 +1,12 @@
 import { after } from 'mocha';
 import { Compiler, ExamplePaths, Features, ALL } from '@tsx-air/types';
 import { loadSuite } from '@tsx-air/testing';
-import { browserify } from '@tsx-air/browserify';
 import { join, basename } from 'path';
 import rimraf from 'rimraf';
 import { preppeteer } from './html/puppeteer.mocha.utils';
 import { packagePath } from '@tsx-air/utils/packages';
-import { safely } from '@tsx-air/utils/src';
+import { safely } from '@tsx-air/utils';
+import { build } from '@tsx-air/builder';
 
 const fixtures = packagePath('@tsx-air/examples', 'fixtures');
 const publicPath = packagePath('@tsx-air/examples', 'public');
@@ -43,7 +43,7 @@ export function shouldCompileExamples(compiler: Compiler, examplePaths: string[]
                             this.timeout(process.env.CI ? 15000 : 6000);
                             this.retries(0);
                             return safely(
-                                () => browserifyFiles(path, paths.temp, compiler),
+                                () => buildTestFiles(path, paths.temp, compiler),
                                 'Failed to compile'
                             );
                         });
@@ -80,8 +80,8 @@ const getUnsupported = (features: Features, compiler: Compiler) => {
     ));
 };
 
-export const browserifyFiles = async (inputDir: string, outputDir: string,
-    compiler: Compiler, entry = 'suite.boilerplate.ts', outputFile = 'boilerplate.js') => await browserify({
+export const buildTestFiles = async (inputDir: string, outputDir: string,
+    compiler: Compiler, entry = 'suite.boilerplate.ts', outputFile = 'boilerplate.js') => await build({
         base: inputDir,
         entry,
         output: join(outputDir, outputFile),
